show discharge model if patient expired while shifting
