+++ src/Components/Facility/models.tsx
@@ -1,6 +1,6 @@
+import { AssignedToObjectModel } from "../Patient/models";
 import { PRNPrescriptionType } from "../Common/prescription-builder/PRNPrescriptionBuilder";
 import { ProcedureType } from "../Common/prescription-builder/ProcedureBuilder";
-import { AssignedToObjectModel } from "../Patient/models";
 
 export interface LocalBodyModel {
   name: string;
@@ -95,10 +95,10 @@ export interface ConsultationModel {
   history_of_present_illness?: string;
   facility?: number;
   facility_name?: string;
-  id?: number;
+  id: string;
   modified_date?: string;
   other_symptoms?: string;
-  patient?: number;
+  patient: string;
   prescribed_medication?: string;
   referred_to?: number | null;
   suggestion?: string;
